Set up common logic files
