+++ f/common_logic/README.md
@@ -1,0 +1,5 @@
+# Common logic
+
+This directory contains common logic that is shared between connector scripts.
+
+For more information on sharing common logic in Windmill, see https://www.windmill.dev/docs/advanced/sharing_common_logic
